project_root/requirements.txt
discord.py>=2.3.0
anthropic>=0.5.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0 